// src/common/enums/user.enum.ts
export enum Role {
  User = "User",
  Admin = "Admin",
}

export enum Provider {
  Google = "google",
  System = "system",
}

export enum Gender {
  Male = "Male",
  Female = "Female",
}
export enum Status {
  Active = "Active",
  Inactive = "Inactive",
  Banned = "Banned",
}

export enum VerificationStatus {
  Pending = "Pending",
  Verified = "Verified",
  Rejected = "Rejected",
}